requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-cors-headers==2.4.0
django-environ==0.4.5
django-oauth-toolkit==1.2.0
django-starterkit==1.0.4
djangorestframework==3.9.0
idna==2.8
oauthlib==3.0.0
Pillow==5.4.1
psycopg2-binary==2.7.6.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
